+++ cmd/command.go
@@ -9,6 +9,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"github.com/giantswarm/devctl/v8/cmd/auth"
 	"github.com/giantswarm/devctl/v8/cmd/completion"
 	"github.com/giantswarm/devctl/v8/cmd/deploy"
 	"github.com/giantswarm/devctl/v8/cmd/gen"
@@ -40,6 +41,19 @@ func New(config Config) (*cobra.Command, error) {
 	}
 
 	var err error
+
+	var authCmd *cobra.Command
+	{
+		c := auth.Config{
+			Stderr: config.Stderr,
+			Stdout: config.Stdout,
+		}
+
+		authCmd, err = auth.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
 
 	var completionCmd *cobra.Command
 	{
@@ -174,6 +188,7 @@ func New(config Config) (*cobra.Command, error) {
 
 	f.Init(c)
 
+	c.AddCommand(authCmd)
 	c.AddCommand(completionCmd)
 	c.AddCommand(deployCmd)
 	c.AddCommand(genCmd)